link login and reset password to home page
