+++ frontend/src/components/ResetPasswordDialog.jsx
@@ -1,0 +1,83 @@
+import React, { useState } from "react";
+import axios from "axios";
+import Dialog from '@material-ui/core/Dialog';
+import TextField from '@material-ui/core/TextField';
+import DialogActions from '@material-ui/core/DialogActions';
+import DialogContent from '@material-ui/core/DialogContent';
+import DialogContentText from '@material-ui/core/DialogContentText';
+import DialogTitle from '@material-ui/core/DialogTitle';
+import Button from '@material-ui/core/Button';
+import { withAuthProps } from '../util/auth';
+import { Link } from "react-router-dom";
+
+function ResetPassword(props) {
+    /* State for handling reset password modal */
+    const [resetDialogOpen, setResetDialogOpen] = useState(false);
+    const [resetEmail, setResetEmail] = useState("");
+
+    async function handleReset(event) {
+        event.preventDefault();
+        // TODO alert?
+        try {
+            const response = await axios.post('/api/accounts/reset_password/request', {
+                "username": resetEmail,
+                "reset_url": window.location.protocol + '//' + window.location.host + '/reset_password'
+            });
+            alert("An email has been sent (if the account exists)!");
+            handleResetClose()
+        } catch {
+            alert("Error requesting password reset");
+        }
+    }
+
+    function handleResetClose() {
+        setResetEmail("");
+        setResetDialogOpen(false);
+    }
+
+    const ResetDialog = (
+        <Dialog
+            open={resetDialogOpen}
+            onClose={handleResetClose}
+        >
+            <DialogTitle id="form-dialog-title">Reset Password</DialogTitle>
+
+            <DialogContent>
+                <DialogContentText>
+                    Reset your password here
+            </DialogContentText>
+                <TextField
+                    autoFocus
+                    margin="dense"
+                    id="name"
+                    label="Email Address"
+                    type="email"
+                    fullWidth
+                    onChange={e => setResetEmail(e.target.value)}
+                />
+            </DialogContent>
+
+            <DialogActions>
+                <Button onClick={handleResetClose} color="primary">
+                    Cancel
+            </Button>
+                <Button onClick={handleReset} color="primary">
+                    Reset
+            </Button>
+            </DialogActions>
+
+        </Dialog >
+    )
+
+    return (
+        <>
+            <Link style={{ textDecoration: 'none' }} onClick={() => setResetDialogOpen(true)} color="inherit">
+                Reset Password
+        </Link >
+            {ResetDialog}
+        </>
+    )
+}
+
+
+export default withAuthProps(ResetPassword);
